+++ include/io.hpp
@@ -8,7 +8,7 @@ Data readTrainImages(size_t imagecount);
 Data readTrainLables(size_t labelcount);
 Data readTestImages(size_t imagecount);
 Data readTestLables(size_t labelcount);
-void writeAccuracy(const float accuracy);
+void writeAccuracy(const double accuracy);
 }  // namespace MNIST
 
-std::vector<float> readAccuracy(const char* filename);
+std::vector<double> readAccuracy(const char* filename);
